pyproject.toml
[tool.poetry]
name = "aidanabackend"
version = "0.1.0"
description = ""
authors = ["daniiar <daniyarflash.m01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
